requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
dj-database-url
psycopg2-binary
gunicorn
Pillow

